중복된 함수 제거 및 github_utils.py 수정
